pyproject.toml
[tool.poetry]
name = "innotter"
version = "0.1.0"
description = ""
authors = ["Alex <aleks.krutsko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
PyJWT = "^2.4.0"
celery = "^5.2.7"
flower = "^1.2.0"
boto3 = "1.24.59"
drf-nested-routers = "^0.93.4"
fastapi = "^0.83.0"
uvicorn = "^0.18.3"
pika = "^1.3.0"
aio-pika = "^8.2.1"
pytest-django = "^4.5.2"
model-bakery = "^1.7.0"
aioboto3 = "^10.1.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pre-commit = "^2.20.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
